requirements.txt
python-telegram-bot==20.7
fastapi>=0.104.0
uvicorn>=0.24.0
sqlite3